pyproject.toml
[project]
name = "securethecyber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-groq>=0.3.2",
    "python-dotenv>=1.1.0",
]
